Drop cross-paths.
It's a shit sandwich no matter how you slice it, but I prefer to have
particular versions of the library bound to particular versions of Scala.

This is further exacerbated by the fact that pom-util is used by sbt-pom-util
and hence SBT, which tends to have more restrictions on the version of Scala
that can be used, but I'm still fine with saying that you have to use pom-util
0.6 if you want to use it with Scala 2.9.2 or 2.10.

When SBT runs atop Scala 2.11, then pom-util 0.7 can be used. This is already
basically the case because of the only-exists-in-2.11 xml module, but I'd
rather just stick to this policy going forward.

Having myriad Scala-version-tagged versions of the pom-util project floating
around is annoying and confuses other build tools.