POMs (sometimes) inherit depends from their parent.
When a POM is a transitive dependency, it appears not to export inherited
depends to its dependents. When a POM is having its own transitive dependencies
computed, dependencies inherited from its parents are considered as part of the
root set.

It would seem to me that a POM should always inherit its parent's depends, but
hey, when has Maven ever opted for sanity?