Fix issue with, harden, fromFile().
In POMs which used '<relativePath />' to "disable" relative path processing in
normal Maven, our code resolved the parent POM as a directory, and fromFile
choked trying to read it.

We now avoid that up front, and nip any other fromFile failure in the bud
rather than pass it on to hapless callers. I really need to familiarize myself
with monadic error propagation...